requirements.txt
# Crypto RAG System - Requirements
# Phase 1: Data Foundation (Current)

# Core Dependencies
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Development
loguru>=0.7.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
ipython>=8.17.0

# Web Interface
streamlit>=1.29.0

# ===== Phase 2 (Uncomment when ready) =====
# # Vector Database
# chromadb>=0.4.22
# 
# # AI/ML
# openai>=1.6.0
# langchain>=0.1.0
# langchain-community>=0.0.10
# sentence-transformers>=2.2.2
# 
# # Advanced ML
# scikit-learn>=1.3.0
# torch>=2.1.0
# xgboost>=2.0.0
# 
# # Multi-Agent Systems  
# crewai>=0.1.0
# langgraph>=0.0.20